move Symfony Analyzers under Symfony namespace
